requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
openai>=1.50.0
google-genai>=1.0.0
jinja2>=3.1.4
